+++ src/Bootstrap/Elements/CheckBox.php
@@ -70,6 +70,16 @@ class CheckBox extends Div
         return $element;
     }
 
+    public function disabled($disabled = true)
+    {
+        $element = clone $this;
+        $element->control = $disabled
+            ? $element->control->attribute('disabled', 'disabled')
+            : $element->control->forgetAttribute('disabled');
+
+        return $element;
+    }
+
     /** @Override */
     protected function assemble()
     {