+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `marvinlabs/laravel-html-bootstrap-4` will be documented in this file.
 
+## 1.0.1 - 2018/04/25
+
+- Fix #26
+- Fix disabled state of radio buttons
+- Fix disabled state of checkboxes
+
 ## 1.0.0 - 2018/04/25
 
 - Let's consider now that the library enters a more mature state. Hello 1.0.0